fix: Align Core ML opt-in with the ABI-safe framework API
The machine learning framework no longer carries a use_coreml_delegate
field on CwtInputModelConfig. That struct crosses the framework boundary
by value, and the SDK and framework are versioned independently, so
growing it broke mixed-version callers. The framework now exposes
-setCoreMLDelegateEnabled: instead, and CwtInputModelConfig is fixed at
five integers.

Resolve the delegate opt-in through that selector at runtime, so a
framework that predates it keeps the existing CPU/XNNPACK path rather
than mismatching the struct layout. Mirror the framework's header
exactly, including the static assertion that pins the layout. The header
is public, so the assertion also guards consumer builds.

Restore the resize and blend helpers to their original file and keep
only the new model-resolution and lazy-mask logic in a same-file
extension. The previous split moved unchanged code and relaxed three
properties to internal without needing to.

Reject non-finite image extents before transforming them, reinitialize
segmentation when the channel count changes, and guard image layout and
data length before the test compares raw bytes.

Testing:
- SDK simulator Debug build
- BackgroundFilterTests and BackgroundFilterCIImageTest: 8 tests, 0 failures
- Header verified byte-identical to the framework's copy

Core ML itself is still exercised only through the fallback branch
locally, because the linked framework predates the selector. The
delegate path needs a framework built from the merged framework commit.
