+++ AmazonChimeSDK/AmazonChimeSDK/audiovideo/video/backgroundfilter/BackgroundFilterProcessor.swift
@@ -16,7 +16,7 @@ import UIKit
 /// top of a background image.
 public class BackgroundFilterProcessor {
     /// Context used for processing and rendering the final output image.
-    let context = CIContext(options: [.cacheIntermediates: false])
+    private let context = CIContext(options: [.cacheIntermediates: false])
 
     /// `CVPixelBufferPool` used to store the final output image.
     private var bufferPool: CVPixelBufferPool?
@@ -27,6 +27,9 @@ public class BackgroundFilterProcessor {
     /// Used to track buffer pool height.
     private var bufferPoolHeight: Int = 0
 
+    /// Used to track the segmentation processor's initialized channel count.
+    private var segmentationProcessorChannels: Int = 0
+
     /// A segmentation processor used to predict foreground of an image.
     /// See `SegmentationProcessor` for more details.
     private let segmentationProcessor: SegmentationProcessor
@@ -35,10 +38,10 @@ public class BackgroundFilterProcessor {
     let logger: Logger
 
     /// Segmentation processor height.
-    var segmentationProcessorHeight = 256
+    private var segmentationProcessorHeight = 256
 
     /// Segmentation processor width.
-    var segmentationProcessorWidth = 144
+    private var segmentationProcessorWidth = 144
 
     /// Static method to check whether BackgroundFilterProcessor can be used. This verifies that the builder
     /// has linked the necessary runtime framework (i.e. `AmazonChimeSDKMachineLearning`) to
@@ -89,6 +92,127 @@ public class BackgroundFilterProcessor {
         return CIImage(cgImage: upscaledMaskImage)
     }
 
+    /// Blends foreground alpha mask with input image to produce a foreground image which is rendered on top
+    /// of a background image using `CIBlendWithAlphaMask` CIFilter.
+    ///
+    /// - Parameters:
+    ///   - inputFrameCI: Input image which is used to blend the foreground alpha mask to produce the foreground image.
+    ///   - maskImage: Foreground alpha mask.
+    ///   - backgroundImage: Background image which can be a blurred or a custom background image.
+    public func blendWithWithAlphaMask(inputFrameCI: CIImage,
+                                       maskImage: CIImage,
+                                       backgroundImage: CIImage) -> CIImage? {
+        guard let blendFilter = CIFilter(name: "CIBlendWithAlphaMask") else {
+            logger.error(msg: "Error creating CIBlendWithAlphaMask CIFilter.")
+            return nil
+        }
+
+        blendFilter.setValue(backgroundImage, forKey: "inputBackgroundImage")
+        blendFilter.setValue(inputFrameCI, forKey: "inputImage")
+        blendFilter.setValue(maskImage, forKey: "inputMaskImage")
+
+        // Create the output image.
+        guard let outputImage = blendFilter.outputImage else {
+            logger.error(msg: "Error creating the blended output image.")
+            return nil
+        }
+
+        return outputImage
+    }
+
+    /// Updates the buffer pool if the previous and new frame dimensions don't match.
+    ///
+    /// - Parameters:
+    ///   - newWidth: New frame width.
+    ///   - newHeight: New frame height.
+    private func updateBufferPool(newWidth: Int, newHeight: Int) {
+        var attributes: [NSString: NSObject] = [:]
+        attributes[kCVPixelBufferPixelFormatTypeKey] = NSNumber(value: Int(kCVPixelFormatType_32BGRA))
+        attributes[kCVPixelBufferWidthKey] = NSNumber(value: newWidth)
+        attributes[kCVPixelBufferHeightKey] = NSNumber(value: newHeight)
+        attributes[kCVPixelBufferIOSurfacePropertiesKey] = [AnyHashable: Any]() as NSObject
+        CVPixelBufferPoolCreate(nil, nil, attributes as CFDictionary?, &bufferPool)
+
+        bufferPoolWidth = newWidth
+        bufferPoolHeight = newHeight
+    }
+
+    /// - Returns: Buffer pool used to store the final image data.
+    public func getBufferPool() -> CVPixelBufferPool? {
+        return bufferPool
+    }
+
+    /// Resize CGImage to the given size.
+    ///
+    /// - Parameters:
+    ///   - image: Input image to resize.
+    ///   - newSize: Size of the image output.
+    ///
+    /// - Returns: Resized image.
+    private func resizeImage(image: CGImage, newSize: CGSize) -> CGImage? {
+        guard let colorSpace = image.colorSpace else {
+            logger.error(msg: "Unable to find color space")
+            return nil
+        }
+
+        let width = Int(newSize.width)
+        let height = Int(newSize.height)
+        let bitsPerComponent = image.bitsPerComponent
+        let bitsPerPixel = image.bitsPerPixel
+
+        // Adjust bitmap info to take into account alpha info.
+        var bitmapInfo = image.bitmapInfo
+        var alphaInfo = bitmapInfo.rawValue & CGBitmapInfo.alphaInfoMask.rawValue
+
+        // Since iOS 8, we cannot create contexts with unmultiplied alpha info.
+        // As a result, if we see a bitmap with an alpha channel that is set to
+        // the last, we assume that the alpha has already been premultiplied and
+        // do nothing. If the alpha channel is set to first, we similarly still
+        // assume it has already been premultiplied and do nothing. We do nothing
+        // because the value should not have been there in the first place and we
+        // prefer doing nothing over modifying the image erroneously.
+        if alphaInfo == CGImageAlphaInfo.last.rawValue {
+            alphaInfo = CGImageAlphaInfo.premultipliedLast.rawValue
+        } else if alphaInfo == CGImageAlphaInfo.first.rawValue {
+            alphaInfo = CGImageAlphaInfo.premultipliedFirst.rawValue
+        }
+
+        // Set the new alpha info in case we have changed it.
+        // In order to set the value, we take the existing alpha
+        // info mask, which is used to extract the value, and invert
+        // it instead. This allows us to instead set the bit value
+        // with the `alphaInfo` bit. As an example, let's say we
+        // have
+        //    bitmapInfo    = 0010
+        //    alphaInfoMask = 0001
+        //    alphaInfo     = 0001
+        // In this case, we'd like to set the bit of `0` to the
+        // least significant bit. Inverting the `alphaInfoMask`
+        // to 1110 keeps all the other bits the same while clearing
+        // out the least significant bit for us to set it. In the case
+        // that `alphaInfo` is one (as in this case), it will be set to
+        // 1. If it was 0, it will also be set to 0 (i.e. no change).
+        bitmapInfo = CGBitmapInfo(rawValue: bitmapInfo.rawValue & ~CGBitmapInfo.alphaInfoMask.rawValue | alphaInfo)
+
+        guard let context = CGContext(data: nil,
+                                      width: width,
+                                      height: height,
+                                      bitsPerComponent: bitsPerComponent,
+                                      bytesPerRow: width * (bitsPerPixel / bitsPerComponent),
+                                      space: colorSpace,
+                                      bitmapInfo: bitmapInfo.rawValue) else {
+            logger.error(msg: "Unable to create context when resizing image")
+            return nil
+        }
+        context.interpolationQuality = .high
+        let rect = CGRect(origin: CGPoint.zero, size: newSize)
+        context.draw(image, in: rect)
+
+        return context.makeImage()
+    }
+}
+
+extension BackgroundFilterProcessor {
     /// Creates a model-resolution foreground mask and lazily maps it to the input extent for video effects.
     func createForegroundAlphaMaskWithLazyUpscale(inputFrameCI: CIImage) -> CIImage? {
         guard BackgroundFilterProcessor.isAvailable(),
@@ -124,6 +248,43 @@ public class BackgroundFilterProcessor {
             .transformed(by: upscaleTransform)
             .transformed(by: originTransform)
             .cropped(to: originalExtent)
+    }
+
+    /// Materializes only the model-resolution input rather than a CPU-readable full-resolution frame.
+    func createModelInputImage(inputFrameCI: CIImage) -> CGImage? {
+        let inputExtent = inputFrameCI.extent
+        guard inputExtent.origin.x.isFinite,
+              inputExtent.origin.y.isFinite,
+              inputExtent.width.isFinite,
+              inputExtent.height.isFinite,
+              inputExtent.width > 0,
+              inputExtent.height > 0 else {
+            logger.error(msg: "Unable to create segmentation input from an invalid extent")
+            return nil
+        }
+
+        let normalizedInput = inputFrameCI.transformed(
+            by: CGAffineTransform(translationX: -inputExtent.origin.x, y: -inputExtent.origin.y)
+        )
+        let modelExtent = CGRect(
+            x: 0,
+            y: 0,
+            width: segmentationProcessorWidth,
+            height: segmentationProcessorHeight
+        )
+        let modelScale = CGAffineTransform(
+            scaleX: modelExtent.width / inputExtent.width,
+            y: modelExtent.height / inputExtent.height
+        )
+        let modelInput = normalizedInput
+            .transformed(by: modelScale, highQualityDownsample: true)
+            .cropped(to: modelExtent)
+
+        guard let modelInputCG = context.createCGImage(modelInput, from: modelExtent) else {
+            logger.error(msg: "Error creating model-resolution CGImage of input frame.")
+            return nil
+        }
+        return modelInputCG
     }
 
     private func createModelResolutionForegroundMask(inputFrameCG: CGImage) -> CGImage? {
@@ -183,9 +344,16 @@ public class BackgroundFilterProcessor {
         let originalWidth = Int(originalSize.width)
         let originalHeight = Int(originalSize.height)
 
-        if bufferPool == nil || originalWidth != bufferPoolWidth || originalHeight != bufferPoolHeight {
+        let shouldUpdateBufferPool =
+            bufferPool == nil ||
+            originalWidth != bufferPoolWidth ||
+            originalHeight != bufferPoolHeight
+        if shouldUpdateBufferPool {
             logger.info(msg: "Updating buffer pool with new sizes: \(originalWidth) x \(originalHeight)")
             updateBufferPool(newWidth: originalWidth, newHeight: originalHeight)
+        }
+
+        if shouldUpdateBufferPool || imageChannels != segmentationProcessorChannels {
             let initializeResult = segmentationProcessor.initialize(
                 segmentationProcessorHeight,
                 width: segmentationProcessorWidth,
@@ -195,6 +363,7 @@ public class BackgroundFilterProcessor {
                 logger.error(msg: "Unable to initialize segmentation processor.")
                 return false
             }
+            segmentationProcessorChannels = imageChannels
         }
 
         if segmentationProcessor.getModelState() != CwtModelState.LOADED.rawValue {
@@ -202,27 +371,5 @@ public class BackgroundFilterProcessor {
             return false
         }
         return true
-    }
-
-    /// Updates the buffer pool if the previous and new frame dimensions don't match.
-    ///
-    /// - Parameters:
-    ///   - newWidth: New frame width.
-    ///   - newHeight: New frame height.
-    private func updateBufferPool(newWidth: Int, newHeight: Int) {
-        var attributes: [NSString: NSObject] = [:]
-        attributes[kCVPixelBufferPixelFormatTypeKey] = NSNumber(value: Int(kCVPixelFormatType_32BGRA))
-        attributes[kCVPixelBufferWidthKey] = NSNumber(value: newWidth)
-        attributes[kCVPixelBufferHeightKey] = NSNumber(value: newHeight)
-        attributes[kCVPixelBufferIOSurfacePropertiesKey] = [AnyHashable: Any]() as NSObject
-        CVPixelBufferPoolCreate(nil, nil, attributes as CFDictionary?, &bufferPool)
-
-        bufferPoolWidth = newWidth
-        bufferPoolHeight = newHeight
-    }
-
-    /// - Returns: Buffer pool used to store the final image data.
-    public func getBufferPool() -> CVPixelBufferPool? {
-        return bufferPool
     }
 }
