+++ AmazonChimeSDK/AmazonChimeSDK/audiovideo/video/backgroundfilter/TensorFlowSegmentationProcessor.m
@@ -11,6 +11,9 @@
 #import "CwtEnum.h"
 #import "CwtTfLiteModel.h"
 
+_Static_assert(sizeof(CwtInputModelConfig) == 5 * sizeof(int),
+               "CwtInputModelConfig is a cross-framework ABI and must remain five integers");
+
 // TensorFlowSegmentationProcessor is a wrapper class around AmazonChimeSDKMachineLearning
 // frameworks's implementation of CwtTfLiteModel, which is a TensorFlow Lite
 // implementation of a segmentation processor. Note that users of this class must
@@ -53,6 +56,28 @@
     return self;
 }
 
+// Enables Core ML delegation when supported by the linked machine learning
+// framework. Older framework versions do not expose this selector and retain
+// their existing CPU inference behavior.
+- (void) enableCoreMLDelegateIfAvailable {
+    SEL selector = NSSelectorFromString(@"setCoreMLDelegateEnabled:");
+    if (![_model respondsToSelector:selector]) {
+        return;
+    }
+
+    NSMethodSignature* methodSignature = [_model methodSignatureForSelector:selector];
+    if (!methodSignature) {
+        return;
+    }
+
+    BOOL enabled = YES;
+    NSInvocation* invocation = [NSInvocation invocationWithMethodSignature:methodSignature];
+    [invocation setTarget:_model];
+    [invocation setSelector:selector];
+    [invocation setArgument:&enabled atIndex:2];
+    [invocation invoke];
+}
+
 // initialize instantiates the model for the segmentation processor.
 // Returns whether able to successfully initialize.
 - (BOOL) initialize:(NSInteger)height width:(NSInteger)width channels:(NSInteger)channels {
@@ -62,6 +87,8 @@
         return NO;
     }
 
+    [self enableCoreMLDelegateIfAvailable];
+
     // Initialize parameters to load model.
     NSString* path = [[NSURL fileURLWithPath:bundle] path];
     CwtInputModelConfig modelConfig = {0};
@@ -70,7 +97,6 @@
     modelConfig.in_channels = (int)channels;
     modelConfig.model_range_min = 0;
     modelConfig.model_range_max = 1;
-    modelConfig.use_coreml_delegate = 1;
 
     // Get the method signature and create an invocation of it.
     SEL selector = NSSelectorFromString(@"loadFile:config:");
